requirements.txt
fastapi
uvicorn[standard]
requests
sentence-transformers
faiss-cpu
openai
pydantic
